Issue #3 Updating the Board struct API to prevent leaking implementation details.
Removed Board::get_mut() as it would have allowed users to change a square's
position without going through the Board AI.

Updated Board::get() to return a copy instead of a reference. This allows decoupling
of the internal implementation.

Renamed Board::replace() to Board::set() so there is parody with Board::get().