Issue #4 Messily sprinting ahead to get all the game unit tests to pass.
This commit implements all of the game logic so the unit tests pass. This includes
adding more unit tests for Game::do_move() as it is a fairly complicated method.

The code in this commit focused on getting the core logic working and is not
elegant or maintainable. Refactoring will come in a future commit.