Issue #4 Implementing the Game::Error enum.
This enumeration is used to report game related errors such as trying to do a
move when the game is over.

This commit also adds a few more tests for various edge cases and updates some
existing tests to print out the game board on failure to help with diagnosing
the failure.