Add example single player game
The single player example is a small but complete CLI application that
demonstrates many features of the library. This example provides a
larger demonstration than is contained in the documentation examples.

Additional changes include ensuring the example is built and checked as
part of the Travis CI integration.

See issue #31.