Replace kcov with tarpaulin for code coverage
The build was using the cargo-travis along with kcov to generate
coverage data and submit it to coveralls. Unfortunately, this stopped
working due to a Rust update (https://github.com/roblabla/cargo-travis/issues/66).
Additionally, it was difficult to run this tool locally.

This commit uses cargo-tarpaulin to generate the test coverage data and
upload it to coveralls. Additionally, developers with Linux can easily
generate coverage reports locally with tarpaulin. The CONTRIBUTING file
is updated with instructions on how to generate coverage reports.

See issue #33.