Improve worst case AI game evaluation time
This commit greatly improves the worst case game evaluation time. This
was accomplished by reducing the number of nodes on the evaluation tree
the AI has to examine.

First, for a new game we know the outcome will always be a cat's game so
there is no need to do any evaluation.

Second, once a branch in the tree has lead to the worst possible outcome
there is no need to look at other children of the node in question.

Together, these changes improve the worst case evaluation time by about 150x.

See issue #29.