Refine the ai_difficulties example
The battle progress display is updated at regular intervals. Previously
it was inconsistently updated depending on the number of games played
and the buffering behavior of the underlying output stream.

The example's introductory text is also updated with more details about
the example and how to add custom difficulties.

See issue #34.