Tweak medium and hard AI difficulties
Using the ai difficulties example, I tweaked the difficulty settings to
to what I think is a good starting place. Playing perfect games, it is
possible to beat hard around 15% of the time.  The medium difficulty
is updated to be more aggressive on picking wining squares ... the
player will have to think a move or two ahead!

See issue #34.